add testes, readme e correções.

+++ src/main/java/br/com/donus/bankbackend/model/Account.java
@@ -16,7 +16,7 @@ public class Account {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private long id;
+	private Integer id;
 	
 	@Column(name = "name")
 	private String name;
